Fix Travis CI install phase
npm appears to have issues with projects that use `peerDependencies` and
a self-referencing `npm link`; one or both of those are causing `npm
install` to remove packages when it should only be installing new
packages.

To work around that, this updates the Travis CI config to run the ESLint
install first to save the override version into `package-lock.json`
before running `npm ci`.